server/requirements.txt
astroid==2.5.2
click==7.1.2
Faker==7.0.1
Flask==1.1.2
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.23
protobuf==3.15.6
pycodestyle==2.7.0
pydocstyle==6.0.0
pylint==2.7.4
python-dateutil==2.8.1
python-dotenv==0.16.0
six==1.15.0
snowballstemmer==2.1.0
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.2
watchdog==2.0.2
Werkzeug==1.0.1
wrapt==1.12.1
